requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
jinja2
kafka-python
websockets
email-validator
pydantic[email]
pydantic
python-multipart
certifi
aiokafka
requests
aiokafka
requests